# docs/openapi.yaml
openapi: 3.1.0
info:
  title: XGuard Testnet API
  version: 0.1.0-alpha.0
  description: >-
    Testnet-only x402 v2 facilitator-compatible safety and routing gateway.
    Base Sepolia is non-billable and mainnet is hard-disabled.
  license:
    name: Apache-2.0
servers:
  - url: https://xguard-testnet.maqamapp.workers.dev
    description: Live Base Sepolia testnet Worker
...
paths:
...
  /verify:
    post:
      tags: [x402]
      summary: Verify an x402 v2 authorization without submitting value
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FacilitatorRequest"
      responses:
        "200":
          description: Verification result, including definitive invalid results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResponse"
        "400":
          description: Malformed or invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResponse"
        "413": { description: Request body exceeded the hard limit }
        "415": { description: Content-Type was not application/json }
        "429": { description: Rate or concurrency limit exceeded }
        "503": { description: Unsupported network or unavailable safe route }
  /settle:
    post:
      tags: [x402]
      summary: Settle once through the selected Base Sepolia facilitator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FacilitatorRequest"
      responses:
        "200":
          description: Final settlement result or a cached final result
          headers:
            X-XGuard-Replayed:
              schema: { type: boolean }
              description: Whether this response came from the durable replay cache
            X-XGuard-Payment-Key:
              schema: { type: string }
              description: Immutable logical authorization identity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettleResponse"
        "400": { description: Malformed or invalid request }
        "409":
          {
            description: Conflicting authorization binding or settlement already in progress,
          }
        "413": { description: Request body exceeded the hard limit }
        "415": { description: Content-Type was not application/json }
        "429": { description: Rate or concurrency limit exceeded }
        "503":
          description: Unsupported network, unavailable route, or ambiguous post-submit outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettleResponse"
components:
  schemas:
...
    FacilitatorRequest:
      type: object
      required: [x402Version, paymentPayload, paymentRequirements]
      additionalProperties: false
      properties:
        x402Version: { const: 2 }
        paymentPayload:
          type: object
          required: [x402Version, accepted, payload]
          additionalProperties: true
          properties:
            x402Version: { const: 2 }
            resource:
              type: object
              additionalProperties: true
            accepted:
              $ref: "#/components/schemas/PaymentRequirements"
            payload:
              type: object
              additionalProperties: true
            extensions:
              type: object
              additionalProperties: true
        paymentRequirements:
          $ref: "#/components/schemas/PaymentRequirements"
    PaymentRequirements:
      type: object
      required: [scheme, network, asset, amount, payTo, maxTimeoutSeconds]
      additionalProperties: true
      properties:
        scheme: { const: exact }
        network: { const: "eip155:84532" }
        asset: { type: string, pattern: "^0x[0-9a-fA-F]{40}$" }
        amount: { type: string, pattern: "^[0-9]+$" }
        payTo: { type: string, pattern: "^0x[0-9a-fA-F]{40}$" }
        maxTimeoutSeconds: { type: integer, minimum: 1 }
        extra:
          type: object
          additionalProperties: true
    VerifyResponse:
      type: object
      required: [isValid]
      properties:
        isValid: { type: boolean }
        invalidReason: { type: string }
        invalidMessage: { type: string }
        payer: { type: string }
    SettleResponse:
      type: object
      required: [success, transaction, network]
      properties:
        success: { type: boolean }
        transaction: { type: string }
        network: { type: string }
        payer: { type: string }
        errorReason: { type: string }
        errorMessage: { type: string }